==> 1_2_LAI_extractor_vdo.ipynb <==
import pandas as pd
import numpy as np
import os
import cv2

DIRECTORY = './images/videos/2020_S'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.h264')]
dataset_list.sort()

for FILENAME in dataset_list:
    vidcap = cv2.VideoCapture('./images/videos/2020_S/' + FILENAME)
    count = 0
    prefix = [FILENAME.split('.')[0].split('_')[0], FILENAME.split('.')[0].split('_')[1]]
    start, end = FILENAME.split('_')[-1].split('.')[0].split('-')
    yy, MM, dd, hh, mm = [start[_:_+2] for _ in range(0, len(start), 2)]
    start_time = f'20{yy}-{MM}-{dd} {hh}:{mm}'
    yy, MM, dd, hh, mm = [end[_:_+2] for _ in range(0, len(end), 2)]
    end_time = f'20{yy}-{MM}-{dd} {hh}:{mm}'
    date_range = pd.date_range(start_time, end_time, freq='10 min')
    
    while True:
        success, image = vidcap.read()
        if not success:
            break

        try:
            folder_date = str(date_range[count]).split(' ')[0]
            output_name = f'{prefix[0]}_{prefix[1]}_{folder_date}'
        except IndexError:
            break

        try:
            os.mkdir(f'./images/2020_S/{output_name}')
        except FileExistsError:
            pass
        
        if (image[..., 0] - image[..., 1]).mean() > 1:
            try:
                cv2.imwrite(f'./images/2020_S/{output_name}/{date_range[count]}.jpg', cv2.resize((image), (128, 128)))
            except IndexError:
                _ = date_range[-1]+pd.Timedelta('10 min')*(count-len(date_range)+1)
                cv2.imwrite(f'./images/2020_S/{output_name}/{_}.jpg', cv2.resize((image), (128, 128)))
        count += 1

    vidcap.release

DIRECTORY = './images/videos/2020_W'
file_list = os.listdir(DIRECTORY)
dataset_list = [file for file in file_list if file.endswith('.h264')]
dataset_list.sort()

for FILENAME in dataset_list:
    vidcap = cv2.VideoCapture('./images/videos/2020_W/' + FILENAME)
    count = 0
    prefix = [FILENAME.split('.')[0].split('_')[0], FILENAME.split('.')[0].split('_')[1]]
    start, end = FILENAME.split('_')[-1].split('.')[0].split('-')
    yy, MM, dd, hh, mm = [start[_:_+2] for _ in range(0, len(start), 2)]
    start_time = f'20{yy}-{MM}-{dd} {hh}:{mm}'
    yy, MM, dd, hh, mm = [end[_:_+2] for _ in range(0, len(end), 2)]
    end_time = f'20{yy}-{MM}-{dd} {hh}:{mm}'
    date_range = pd.date_range(start_time, end_time, freq='10 min')
    
    while True:
        success, image = vidcap.read()
        if not success:
            break

        try:
            folder_date = str(date_range[count]).split(' ')[0]
            output_name = f'{prefix[0]}_{prefix[1]}_{folder_date}'
        except IndexError:
            break

        try:
            os.mkdir(f'./images/2020_W/{output_name}')
        except FileExistsError:
            pass
        
        if (image[..., 0] - image[..., 1]).mean() > 1:
            try:
                cv2.imwrite(f'./images/2020_W/{output_name}/{date_range[count]}.jpg', cv2.resize(image, (128, 128)))
            except IndexError:
                _ = date_range[-1]+pd.Timedelta('10 min')*(count-len(date_range)+1)
                cv2.imwrite(f'./images/2020_W/{output_name}/{_}.jpg', cv2.resize(image, (128, 128)))
        count += 1

    vidcap.release

# for FILENAME in dataset_list:
#     vidcap = cv2.VideoCapture('./images/videos/2020_S/' + FILENAME)
#     count = 0
#     output_name = FILENAME.split('.')[0]
#     try:
#         os.mkdir(f'./images/2020_S/{output_name}')
#     except FileExistsError:
#         pass
    
#     start, end = FILENAME.split('_')[-1].split('.')[0].split('-')
#     yy, MM, dd, hh, mm = [start[_:_+2] for _ in range(0, len(start), 2)]
#     start_time = f'20{yy}-{MM}-{dd} {hh}:{mm}'
#     yy, MM, dd, hh, mm = [end[_:_+2] for _ in range(0, len(end), 2)]
#     end_time = f'20{yy}-{MM}-{dd} {hh}:{mm}'
#     date_range = pd.date_range(start_time, end_time, freq='10 min')
    
#     while True:
#         success, image = vidcap.read()
#         if not success:
#             break
#         try:
#             cv2.imwrite(f'./images/2020_S/{output_name}/{date_range[count]}.jpg', image[284:-284, 704:-704, :])
#         except IndexError:
#             _ = date_range[-1]+pd.Timedelta('10 min')*(count-len(date_range)+1)
#             cv2.imwrite(f'./images/2020_S/{output_name}/{_}.jpg', image[284:-284, 704:-704, :])
#         count += 1

#     vidcap.release
