==> 3_weight_input_processing.ipynb <==
import os

import numpy as np
import pandas as pd

# Fresh weights

### 2020 Summer

input_df = pd.read_csv('./results/2020_S/SW2_greenhouse.csv', index_col='Unnamed: 0').interpolate()
input_df.index = pd.DatetimeIndex(input_df.index)

input_df = input_df.resample('10 min').mean()

output_df = pd.read_csv('./results/2020_S/fw_labels.csv', index_col='Unnamed: 0')

input_data_2020S_CT = [pd.concat([input_df.loc[:, :'permit'], input_df.loc[:, 'loadcell_1']], axis=1).values,
                       pd.concat([input_df.loc[:, :'permit'], input_df.loc[:, 'loadcell_2']], axis=1).values,
                       pd.concat([input_df.loc[:, :'permit'], input_df.loc[:, 'loadcell_3']], axis=1).values]

data_indices_2020S_CT = [np.array(output_df['CT_1'].index),
                         np.array(output_df['CT_2'].index),
                         np.array(output_df['CT_3'].index)]

output_label_2020S_CT = [output_df['CT_1'].values,
                         output_df['CT_2'].values,
                         output_df['CT_3'].values]

input_data_2020S = np.concatenate(input_data_2020S_CT)
data_indices_2020S = np.concatenate(data_indices_2020S_CT).reshape(-1, 1)
output_label_2020S = np.concatenate(output_label_2020S_CT).reshape(-1, 1)

### 2020 Winter

input_df = pd.read_csv('./results/2020_W/SW_CT_greenhouse.csv', index_col='Unnamed: 0').interpolate()
input_df.index = pd.DatetimeIndex(input_df.index)

input_df = input_df.resample('10 min').mean()

input_df.loc[:, 'loadcell_1'] = input_df.loc[:, 'loadcell_1'] - 0.8
input_df.loc[:, 'loadcell_2'] = input_df.loc[:, 'loadcell_2'] - 0.5
input_df.loc[:, 'loadcell_3'] = input_df.loc[:, 'loadcell_3'] - 0.8

output_df = pd.read_csv('./results/2020_W/fw_ct_labels.csv', index_col='Unnamed: 0')

input_df = input_df.interpolate()

input_data_2020W_CT = [pd.concat([input_df.loc[:, :'permit'], input_df.loc[:, 'loadcell_1']], axis=1).values,
                       pd.concat([input_df.loc[:, :'permit'], input_df.loc[:, 'loadcell_2']], axis=1).values,
                       pd.concat([input_df.loc[:, :'permit'], input_df.loc[:, 'loadcell_3']], axis=1).values]

data_indices_2020W_CT = [np.array(output_df['CT_1'].index),
                         np.array(output_df['CT_2'].index),
                         np.array(output_df['CT_3'].index)]

output_label_2020W_CT = [output_df['CT_1'].values,
                         output_df['CT_2'].values,
                         output_df['CT_3'].values]

input_data_2020W = np.concatenate(input_data_2020W_CT)
data_indices_2020W = np.concatenate(data_indices_2020W_CT).reshape(-1, 1)
output_label_2020W = np.concatenate(output_label_2020W_CT).reshape(-1, 1)

### Data minmax

cal_input_minmax = np.concatenate([input_data_2020S, input_data_2020W])
cal_output_minmax = np.concatenate([output_label_2020S, output_label_2020W])

INPUT_MAXS = cal_input_minmax.reshape(-1, 9).max(axis=0)
INPUT_MINS = cal_input_minmax.reshape(-1, 9).min(axis=0)
OUTPUT_MAX = cal_output_minmax.max()
OUTPUT_MIN = cal_output_minmax.min()

### Data save

input_data = np.concatenate(input_data_2020S_CT)
data_indices = np.concatenate(data_indices_2020S_CT).reshape(-1, 1)
output_label = np.concatenate(output_label_2020S_CT).reshape(-1, 1)

input_data = (input_data - INPUT_MINS)/(INPUT_MAXS - INPUT_MINS)
input_data = input_data.reshape(-1, 144, 9)
output_label = (output_label - OUTPUT_MIN)/(OUTPUT_MAX - OUTPUT_MIN)

f = open('./results/2020_S/fw_ct_dataset.npz', 'wb')
np.savez(f,
         data_indices = data_indices,
         input_data = input_data,
         output_label = output_label,
         INPUT_MAXS = INPUT_MAXS,
         INPUT_MINS = INPUT_MINS,
         OUTPUT_MAX = OUTPUT_MAX,
         OUTPUT_MIN = OUTPUT_MIN
        )
f.close()

print(input_data.shape)
print(output_label.shape)
print(data_indices.shape)

input_data = np.concatenate(input_data_2020W_CT)
data_indices = np.concatenate(data_indices_2020W_CT).reshape(-1, 1)
output_label = np.concatenate(output_label_2020W_CT).reshape(-1, 1)

input_data = (input_data - INPUT_MINS)/(INPUT_MAXS - INPUT_MINS)
input_data = input_data.reshape(-1, 144, 9)
output_label = (output_label - OUTPUT_MIN)/(OUTPUT_MAX - OUTPUT_MIN)

f = open('./results/2020_W/fw_ct_dataset.npz', 'wb')
np.savez(f,
         data_indices = data_indices,
         input_data = input_data,
         output_label = output_label,
         INPUT_MAXS = INPUT_MAXS,
         INPUT_MINS = INPUT_MINS,
         OUTPUT_MAX = OUTPUT_MAX,
         OUTPUT_MIN = OUTPUT_MIN
        )
f.close()

print(input_data.shape)
print(output_label.shape)
print(data_indices.shape)
