==> 4_PBM_SanMol.ipynb <==
import sys
sys.path.append('../')

import numpy as np
import pandas as pd

from SanMol.growth_model import *

# Parameters

START_DATE = '2020-02-26'
END_DATE = '2020-07-03'

START_DATE = pd.Timestamp(START_DATE)
END_DATE = pd.Timestamp(END_DATE)

print(f'Cultivation period: {(END_DATE - START_DATE).days} days')
print(f'Starting DOY: {START_DATE.dayofyear}')

env_df = pd.read_csv('../ch2_weighing/results/2020_S/SW_CT_greenhouse_origin.csv', index_col='Unnamed: 0')
env_df['co2'] = 400
env_df.index = pd.DatetimeIndex(env_df.index)
env_df = env_df.loc[:, ['temp', 'hum', 'co2', 'rad']]

env_df = env_df.resample('1H').mean()

crop = SweetPepper()

temp = env_df.values[:24, 0].reshape(-1, 1)
rh = env_df.values[:24, 1].reshape(-1, 1)/100
co2 = env_df.values[:24, 2].reshape(-1, 1)
rad = env_df.values[:24, 3].reshape(-1, 1)

for i in range(150):
    crop.life_cycle(temp, rh, co2, rad)

# crop.life_cycle(temp, rh, co2, rad)

crop.status()
